Rename IdentifierVariable to AlgebraicExpressionVariable
Also move variables code to a new file `variables.go` instead of
`stack.go`
